cache testcase assertions can now be called with a cache_alias to force operation on a cache other than the default for the active model
cache testcase helper 'populate_cache' can now be supplied with extra caches to populate, in the event that keys in non-base caches need to be checked

new cache mode 'asset' allows caching of sub-items under a parent item. If 'asset' mode is enabled, a detail_version key will be maintained, and used as the basis for cached asset keys. Updating or deleting the parent will invalidate all child asset keys